+++ packages/constants/src/index.ts
@@ -1,1 +1,2 @@
 export * from './URL'
+export * from './regex'
