+++ packages/utils/src/index.ts
@@ -1,1 +1,2 @@
 export * from './iframeEventHandler'
+export * from './format'
