+++ web/src/components/auth/SignupModal/AccountForm.tsx
@@ -1,3 +1,4 @@
+import { PASSWORD_REGEX } from '@allwagelab/constants'
 import { Button } from '@allwagelab/react'
 import styled from '@emotion/styled'
 import { zodResolver } from '@hookform/resolvers/zod'
@@ -9,7 +10,6 @@ import { z } from 'zod'
 import type { SignupFormData } from '@/types/auth'
 
 import { checkEmailDuplicate } from '@/apis/auth'
-import { PASSWORD_REGEX } from '@/lib/constants'
 import {
   ErrorMessage,
   Form,
