+++ web/src/mocks/auth.ts
@@ -1,6 +1,5 @@
+import { PHONE_NUMBER_REGEX } from '@allwagelab/constants'
 import { http, HttpResponse, type PathParams } from 'msw'
-
-import { PHONE_NUMBER_REGEX } from '@/lib/constants'
 
 const createURL = (url: string) => {
   const baseURL = import.meta.env.VITE_BASE_URL
